Move pytest riak helper into testutils
